Refactor groups functionality into its own service
