+++ src/app/issues-viewer/issues-viewer.component.ts
@@ -8,6 +8,7 @@ import { ErrorMessageService } from '../core/services/error-message.service';
 import { FiltersService } from '../core/services/filters.service';
 import { GithubService } from '../core/services/github.service';
 import { GroupingContextService } from '../core/services/grouping/grouping-context.service';
+import { GroupService } from '../core/services/grouping/groups.service';
 import { IssueService } from '../core/services/issue.service';
 import { LabelService } from '../core/services/label.service';
 import { MilestoneService } from '../core/services/milestone.service';
@@ -37,12 +38,6 @@ export class IssuesViewerComponent implements OnInit, AfterViewInit, OnDestroy {
 
   popStateNavigationId: number;
 
-  /** Users to show as columns */
-  groups: Group[] = [];
-
-  /** The list of users with 0 issues (hidden) */
-  hiddenGroups: Group[] = [];
-
   @ViewChildren(CardViewComponent) cardViews: QueryList<CardViewComponent>;
 
   views = new BehaviorSubject<QueryList<CardViewComponent>>(undefined);
@@ -53,6 +48,7 @@ export class IssuesViewerComponent implements OnInit, AfterViewInit, OnDestroy {
     public issueService: IssueService,
     public labelService: LabelService,
     public milestoneService: MilestoneService,
+    public groupService: GroupService,
     public groupingContextService: GroupingContextService,
     private router: Router,
     private filtersService: FiltersService
@@ -106,50 +102,9 @@ export class IssuesViewerComponent implements OnInit, AfterViewInit, OnDestroy {
       this.availableGroupsSubscription.unsubscribe();
     }
 
-    // Fetch assignees
-    this.groups = [];
-    this.hiddenGroups = [];
+    this.groupService.groups = [];
+    this.groupService.hiddenGroups = [];
 
-    this.availableGroupsSubscription = this.groupingContextService.getGroups().subscribe((x) => (this.groups = x));
-  }
-
-  /**
-   * Checks if our current repository available on view service is indeed a valid repository
-   */
-  private checkIfValidRepository() {
-    const currentRepo = this.viewService.currentRepo;
-
-    if (Repo.isInvalidRepoName(currentRepo)) {
-      return of(false);
-    }
-
-    return this.githubService.isRepositoryPresent(currentRepo.owner, currentRepo.name);
-  }
-
-  /**
-   * Update the list of hidden group based on the new info.
-   * @param issueLength The number of issues under this group.
-   * @param target The group.
-   */
-  updateHiddenGroups(issueLength: number, target: Group) {
-    // If the group is in the hidden list, add it if it has no issues.
-    // Also add it if it is unchecked in the filter.
-    if (issueLength === 0 && this.groupingContextService.isInHiddenList(target)) {
-      this.addToHiddenGroups(target);
-    } else {
-      this.removeFromHiddenGroups(target);
-    }
-  }
-
-  private addToHiddenGroups(target: Group) {
-    const isGroupPresent = this.hiddenGroups.some((group) => group.equals(target));
-
-    if (!isGroupPresent) {
-      this.hiddenGroups.push(target);
-    }
-  }
-
-  private removeFromHiddenGroups(target: Group) {
-    this.hiddenGroups = this.hiddenGroups.filter((group) => !group.equals(target));
+    this.availableGroupsSubscription = this.groupingContextService.getGroups().subscribe((x) => (this.groupService.groups = x));
   }
 }
